% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug.similarity-package.R
\docType{package}
\name{drug.similarity-package}
\alias{drug.similarity}
\alias{drug.similarity-package}
\title{drug.similarity: Compute several drug similarity metrics}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Clemens Hug \email{clemens.hug@gmail.com} (\href{https://orcid.org/0000-0002-8299-3274}{ORCID})

}
\keyword{internal}
