% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms.R
\name{getInfo}
\alias{getInfo}
\alias{getInfo,microsimulation-generic}
\title{getInfo Generic}
\usage{
getInfo(object)
}
\arguments{
\item{object}{A microsimulation object to get its information.}
}
\description{
getInfo Generic
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}

