% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms.R
\docType{methods}
\name{getTAEs,microsimulation-method}
\alias{getTAEs,microsimulation-method}
\title{getTAEs Method}
\usage{
\S4method{getTAEs}{microsimulation}(object)
}
\arguments{
\item{object}{A microsimulation object to get its information.}
}
\value{
taes A list of numbers indicating the Total Absolute Error of the fitting process 
for each of the census areas.
}
\description{
getTAEs Method
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}

