\name{getTAEs}
\alias{getTAEs}
\alias{getTAEs,microsimulation-generic}
\alias{getTAEs,microsimulation-method}
\title{getTAEs Generic}
\arguments{
  \item{object}{A microsimulation object to get its
  information.}

  \item{object}{A microsimulation object to get its
  information.}
}
\value{
  taes A list of numbers indicating the Total Absolute
  Error of the fitting process for each of the census
  areas.
}
\description{
  Get the TAE from a microsimulation object.

  getTAEs Method
}
\author{
  Dimitris Kavroudakis \email{dimitris123@gmail.com}

  Dimitris Kavroudakis \email{dimitris123@gmail.com}
}

