% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms.R
\name{mysetSeed}
\alias{mysetSeed}
\title{mysetSeed}
\usage{
mysetSeed(inseed)
}
\arguments{
\item{inseed}{A number to set as a random seed.}
}
\description{
mysetSeed
}
\details{
mysetSeed
}
\examples{
library(sms)
sms::mysetSeed(1900)
}

