% Generated by roxygen2 (4.0.1): do not edit by hand
\name{random_panel_selection}
\alias{random_panel_selection}
\title{random_panel_selection}
\usage{
random_panel_selection(indf, n)
}
\arguments{
\item{indf}{The initial dataframe from wich a selection will be made.}

\item{n}{The number of random rows}
}
\value{
a selection of rows as a dataframe
}
\description{
Select n random rows from a dataframe
}
\details{
Select n random rows from a dataframe
}
\examples{
library(sms)
data(survey) #load the data
data(census)

some.individuals=random_panel_selection(survey,4)
print(some.individuals)     # Print the selection of individuals
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}

