% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms.R
\docType{package}
\name{sms-package}
\alias{sms-package}
\title{Spatial Microsimulation Library}
\description{
Generate small area population microdata from census and survey datasets. 
Fit the survey data to census area descriptions and export the population of small areas (microdata).
}
\details{
Generate small area population microdata from census and panel datasets. 
Fit the survey data to census area descriptions and export the popultion of small areas.
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}
\references{
Dimitris Kavroudakis D (2015). \strong{sms: An R Package for the Construction of Microdata for 
Geographical Analysis.} \emph{Journal of Statistical Software}, \strong{68}(2), pp. 1-23. \url{http://10.18637/jss.v068.i02}
}

