% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{sms_compound_ids}
\alias{sms_compound_ids}
\title{Convert compound names to SMS compound IDs (lspci_id)}
\usage{
sms_compound_ids(ids)
}
\value{
A named vector with lspci_ids as values and corresponding queries
as names
}
\description{
Convert compound names to SMS compound IDs (lspci_id)
}
