% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_delete_app}
\alias{sms_delete_app}
\title{Deletes an existing application}
\usage{
sms_delete_app(appId, forceStopAppReplication, forceTerminateApp)
}
\arguments{
\item{appId}{ID of the application to delete.}

\item{forceStopAppReplication}{While deleting the application, stop all replication jobs corresponding
to the servers in the application.}

\item{forceTerminateApp}{While deleting the application, terminate the stack corresponding to the
application.}
}
\description{
Deletes an existing application. Optionally deletes the launched stack
associated with the application and all AWS SMS replication jobs for
servers in the application.
}
\section{Request syntax}{
\preformatted{sms$delete_app(
  appId = "string",
  forceStopAppReplication = TRUE|FALSE,
  forceTerminateApp = TRUE|FALSE
)
}
}

\keyword{internal}
