% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_generate_change_set}
\alias{sms_generate_change_set}
\title{Generates a target change set for a currently launched stack and writes
it to an Amazon S3 object in the customer’s Amazon S3 bucket}
\usage{
sms_generate_change_set(appId, changesetFormat)
}
\arguments{
\item{appId}{ID of the application associated with the change set.}

\item{changesetFormat}{Format for the change set.}
}
\description{
Generates a target change set for a currently launched stack and writes
it to an Amazon S3 object in the customer's Amazon S3 bucket.
}
\section{Request syntax}{
\preformatted{sms$generate_change_set(
  appId = "string",
  changesetFormat = "JSON"|"YAML"
)
}
}

\keyword{internal}
