% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_app_replication_configuration}
\alias{sms_get_app_replication_configuration}
\title{Retrieves an application replication configuration associatd with an
application}
\usage{
sms_get_app_replication_configuration(appId)
}
\arguments{
\item{appId}{ID of the application associated with the replication configuration.}
}
\description{
Retrieves an application replication configuration associatd with an
application.
}
\section{Request syntax}{
\preformatted{sms$get_app_replication_configuration(
  appId = "string"
)
}
}

\keyword{internal}
