% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_get_connectors}
\alias{sms_get_connectors}
\title{Describes the connectors registered with the AWS SMS}
\usage{
sms_get_connectors(nextToken, maxResults)
}
\arguments{
\item{nextToken}{The token for the next set of results.}

\item{maxResults}{The maximum number of results to return in a single call. The default
value is 50. To retrieve the remaining results, make another call with
the returned \code{NextToken} value.}
}
\description{
Describes the connectors registered with the AWS SMS.
}
\section{Request syntax}{
\preformatted{sms$get_connectors(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
