% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{sms_phenotypic_similarity}
\alias{sms_phenotypic_similarity}
\title{Calculate phenotypic similarity}
\usage{
sms_phenotypic_similarity(
  query_ids,
  target_ids = NULL,
  min_n = 4,
  show_compound_names = FALSE
)
}
\arguments{
\item{query_ids}{Compounds (names or \code{lspci_ids}) for which to calculate
similarity. If no targets are provided, the pairwise similarity of
the query compounds with all other compounds is calculated.}

\item{target_ids}{If provided (names or \code{lspci_ids}), all pairwise
similarities between query and target compounds are calculated.}

\item{min_n}{Minimum number of shared assays between compounds}

\item{show_compound_names}{Add columns with compound names to output}
}
\value{
A 3 column data frame with query_id, target_id, and similarity
}
\description{
Computes similarity between compounds based on phenotypic assays performed
on them that are published in ChEMBL. The similarity metric is the Pearson
correlation between the normalized results of all assays shared by both
compounds.
}
\details{
Before similarity functions can be used compound data needs
to be downloaded to the working directory from the
\href{https://www.synapse.org/#!Synapse:syn25955270}{Synapse repository} or
automatically using the \code{\link[=sms_download]{sms_download()}} function.

Compound names can be provided directly to the function,
mapped to \code{lspci_ids} using the \code{\link[=sms_compound_ids]{sms_compound_ids()}}
function, or using the \code{lsp_compound_names} table from
\url{https://labsyspharm.shinyapps.io/smallmoleculesuite}.
}
\examples{
sms_phenotypic_similarity(c("ruxolitinib", "tofacitinib"), show_compound_names = TRUE)

}
\seealso{
\code{\link[=sms_download]{sms_download()}} for downloading SMS data

Other similarity functions: 
\code{\link{sms_all_similarities}()},
\code{\link{sms_chemical_similarity}()},
\code{\link{sms_tas_similarity}()}
}
\concept{similarity functions}
