% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms_operations.R
\name{sms_put_app_replication_configuration}
\alias{sms_put_app_replication_configuration}
\title{Creates or updates a replication configuration for an application}
\usage{
sms_put_app_replication_configuration(appId,
  serverGroupReplicationConfigurations)
}
\arguments{
\item{appId}{ID of the application tassociated with the replication configuration.}

\item{serverGroupReplicationConfigurations}{Replication configurations for server groups in the application.}
}
\description{
Creates or updates a replication configuration for an application.
}
\section{Request syntax}{
\preformatted{sms$put_app_replication_configuration(
  appId = "string",
  serverGroupReplicationConfigurations = list(
    list(
      serverGroupId = "string",
      serverReplicationConfigurations = list(
        list(
          server = list(
            serverId = "string",
            serverType = "VIRTUAL_MACHINE",
            vmServer = list(
              vmServerAddress = list(
                vmManagerId = "string",
                vmId = "string"
              ),
              vmName = "string",
              vmManagerName = "string",
              vmManagerType = "VSPHERE"|"SCVMM"|"HYPERV-MANAGER",
              vmPath = "string"
            ),
            replicationJobId = "string",
            replicationJobTerminated = TRUE|FALSE
          ),
          serverReplicationParameters = list(
            seedTime = as.POSIXct(
              "2015-01-01"
            ),
            frequency = 123,
            runOnce = TRUE|FALSE,
            licenseType = "AWS"|"BYOL",
            numberOfRecentAmisToKeep = 123,
            encrypted = TRUE|FALSE,
            kmsKeyId = "string"
          )
        )
      )
    )
  )
)
}
}

\keyword{internal}
