\docType{data}
\name{survey}
\alias{survey}
\title{A survey dataset of 200 individuals}
\format{A data frame with 200 rows and 7 variables}
\usage{
data(survey)
}
\description{
A sample survey dataset containing binary (0 or 1)
information about 200 individuals. Those individuals will
be used to populate the simulated areas.  The variables in
the dataset are as follows: \itemize{ \item pid: The unique
indentifier of the individual \item female: Binary value of
the sex of the individual. 1-Female, 0-Male \item
agemature: Binary value indicating if the individual
belongs to the mature age group.  0-No, 1-Yes \item
car_owner: Binary value indicating if the individual owns a
car. 0-No, 1-Yes \item house_owner: Binary value indicating
if the individual owns a house. 0-No, 1-Yes \item working:
Binary value indicating if the individual is working. 0-No,
1-Yes }
}
\keyword{datasets}

