\name{getInfo}
\alias{getInfo}
\alias{getInfo,microsimulation-generic}
\alias{getInfo,microsimulation-method}
\title{getInfo Generic}
\arguments{
  \item{object}{A microsimulation object to get its
  information.}

  \item{object}{A microsimulation object to get its
  information.}
}
\description{
  getInfo Generic

  Get information from a microsimulation object
}
\author{
  Dimitris Kavroudakis \email{dimitris123@gmail.com}

  Dimitris Kavroudakis \email{dimitris123@gmail.com}
}

