% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getTAEs}
\alias{getTAEs}
\alias{getTAEs,microsimulation-generic}
\title{getTAEs Generic}
\usage{
getTAEs(object)
}
\arguments{
\item{object}{A microsimulation object to get its information.}
}
\description{
Get the TAE from a microsimulation object.
}
\author{
Dimitris Kavroudakis \email{dimitris123@gmail.com}
}

