\docType{package}
\name{sms-package}
\alias{sms-package}
\title{Spatial Microsimulation Library}
\description{
  Generate small area population microdata from census and
  survey datasets. Fit the survey data to census area
  descriptions and export the population of small areas.
}
\details{
  Please feel free to contact me for error, problems or
  other features of the library.
  \email{dimitris123@gmail.com}

  Generate small area population microdata from census and
  panel datasets. Fit the survey data to census area
  descriptions and export the popultion of small areas.
}
\author{
  Dimitris Kavroudakis \email{dimitris123@gmail.com}
}

