\docType{package}
\name{sms-package}
\alias{sms-package}
\title{Spatial Microsimulation Library}
\description{
  Generate small area population microdata from census and
  panel datasets. Fit the panel data to census area
  descriptions.
}
\details{
  Small area population estimates.
}
\author{
  Dimitris Kavroudakis \email{dimitris123@gmail.com}
}

